requirements.txt
fhir.resources==7.1.0
pytest==8.2.2
pytest-mock==3.14.0
pytest-cov==4.1.0
pytest-asyncio==0.26.0
ruff==0.9.1
usdm4>=0.7.0
